Improve README
There are two main use cases for this repo. The first is building the
oidc-testprovider Docker image which helps teams do local development by
acting as an OIDC OP. The second is building a bunch of images used by
mozilla-django-oidc and other things for e2e tests.

Since the first is the one used by other people, it makes sense to focus
on documentation for that so it's easier for people to use.

This also updates image names.