Add oidc_e2e_setup_py for each supported Python version
Previously, we had an oidc_e2e_setup_py2 and an oidc_e2e_setup_py3 which
sufficed for our testing needs. After we dropped support for Python 2.7,
we dropped the oidc_e2e_setup_py2.

However, Django 4 requires Python >= 3.8 and we want to have a Python
3.7 image.

This changes our support a bit so now we're going to have an image for
every supported Python version. This adds the missing images and updates
the documentation accordingly.

This also updates the Python images used to be pinned to a specific
Python version (e.g. 3.7.13) and also update them to use bullseye
(current Debian stable) since stretch is not longer supported.